+++ CHANGELOG.md
@@ -6,6 +6,10 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 
 ## [Unreleased]
 
+### Fixed
+
+- Italian translation by [@svityboy](https://gitlab.gnome.org/svityboy).
+
 ## [22.0.0 Millet] - 2022-09-10
 
 ### Fixed